requirements.txt
gunicorn==19.7.1
Django==1.11.9
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
pytz==2017.3
